refactor(phase-4): move git, github, source-parser to @agents/core
Move 3 integration modules from packages/cli/src/lib/ to
packages/core/src/ and update all imports (static + dynamic + mocks).

Remove leaked runtime mock from external-skills-refactor.test.ts
to prevent cross-file contamination with package-scoped mock.module.
Add @napi-rs/keyring and @octokit deps to core package.json.

Tests: 1882 pass / 50 fail / 1 error (zero regressions from Phase 3)
