+++ packages/cli/src/lib/external-skills.ts
@@ -12,6 +12,8 @@
 import { existsSync, mkdirSync, readdirSync, readFileSync, rmSync, writeFileSync } from 'node:fs'
 import { tmpdir } from 'node:os'
 import { join, relative } from 'node:path'
+import { cloneRepo, gitRaw, lsRemote } from '@agents/core/git'
+import { addComment, createIssue, type Issue, searchIssues } from '@agents/core/github'
 import { formatHash, hashDirectory, lockKey } from '@agents/core/hash'
 import { spawnSync } from '@agents/core/runtime'
 import {
@@ -23,8 +25,6 @@ import { checkSymlink, createSymlink } from '@agents/core/symlink'
 import { CliError, err, ok, type Result } from '@agents/core/types'
 import yaml from 'js-yaml'
 import * as v from 'valibot'
-import { cloneRepo, gitRaw, lsRemote } from './git'
-import { addComment, createIssue, type Issue, searchIssues } from './github'
 
 // ---------------------------------------------------------------------------
 // Types
