+++ packages/cli/src/lib/skill-outdated.ts
@@ -13,10 +13,10 @@
  */
 
 import { join } from 'node:path'
+import { fetchSkillFolderHash, lsRemote } from '@agents/core/git'
 import { hashDirectory } from '@agents/core/hash'
 import type { LockfileV1, SkillLockEntry } from '@agents/core/schemas'
 import { CliError, err, ok, type Result } from '@agents/core/types'
-import { fetchSkillFolderHash, lsRemote } from './git'
 import { readLockfile } from './lockfile'
 
 // ---------------------------------------------------------------------------
