+++ packages/cli/test/skill-outdated.test.ts
@@ -53,7 +53,7 @@ describe('github source', () => {
   test('reports "current" when fetchSkillFolderHash returns matching hash', async () => {
     const HASH = 'a'.repeat(64)
 
-    mock.module('../src/lib/git', () => ({
+    mock.module('@agents/core/git', () => ({
       fetchSkillFolderHash: async () => HASH,
       lsRemote: async () => ({ ok: true, value: 'unused' }),
     }))
@@ -76,7 +76,7 @@ describe('github source', () => {
     const LOCAL_HASH = 'a'.repeat(64)
     const REMOTE_HASH = 'b'.repeat(64)
 
-    mock.module('../src/lib/git', () => ({
+    mock.module('@agents/core/git', () => ({
       fetchSkillFolderHash: async () => REMOTE_HASH,
       lsRemote: async () => ({ ok: true, value: 'unused' }),
     }))
@@ -95,7 +95,7 @@ describe('github source', () => {
   })
 
   test('reports "unavailable" when Trees API returns null', async () => {
-    mock.module('../src/lib/git', () => ({
+    mock.module('@agents/core/git', () => ({
       fetchSkillFolderHash: async () => null,
       lsRemote: async () => ({ ok: true, value: 'unused' }),
     }))
@@ -126,7 +126,7 @@ describe('local source', () => {
     const realHash = await hashDirectory(join(tmp, 'content', 'skills', 'local-skill'))
 
     // No need to mock hash module for local -- use real implementation
-    mock.module('../src/lib/git', () => ({
+    mock.module('@agents/core/git', () => ({
       fetchSkillFolderHash: async () => null,
       lsRemote: async () => ({ ok: false, error: { message: 'unused' } }),
     }))
@@ -146,7 +146,7 @@ describe('local source', () => {
   test('reports "outdated" when directory hash differs', async () => {
     await createSkillDir(tmp, 'changed-skill', '# Changed content')
 
-    mock.module('../src/lib/git', () => ({
+    mock.module('@agents/core/git', () => ({
       fetchSkillFolderHash: async () => null,
       lsRemote: async () => ({ ok: false, error: { message: 'unused' } }),
     }))
@@ -167,7 +167,7 @@ describe('local source', () => {
   })
 
   test('reports "unavailable" when directory is missing', async () => {
-    mock.module('../src/lib/git', () => ({
+    mock.module('@agents/core/git', () => ({
       fetchSkillFolderHash: async () => null,
       lsRemote: async () => ({ ok: false, error: { message: 'unused' } }),
     }))
@@ -195,7 +195,7 @@ describe('local source', () => {
 
 describe('edge cases', () => {
   test('empty lock file returns empty array', async () => {
-    mock.module('../src/lib/git', () => ({
+    mock.module('@agents/core/git', () => ({
       fetchSkillFolderHash: async () => null,
       lsRemote: async () => ({ ok: false, error: { message: 'unused' } }),
     }))
@@ -209,7 +209,7 @@ describe('edge cases', () => {
   })
 
   test('no lock file returns empty array', async () => {
-    mock.module('../src/lib/git', () => ({
+    mock.module('@agents/core/git', () => ({
       fetchSkillFolderHash: async () => null,
       lsRemote: async () => ({ ok: false, error: { message: 'unused' } }),
     }))
@@ -222,7 +222,7 @@ describe('edge cases', () => {
   })
 
   test('unknown source type returns "unknown" status', async () => {
-    mock.module('../src/lib/git', () => ({
+    mock.module('@agents/core/git', () => ({
       fetchSkillFolderHash: async () => null,
       lsRemote: async () => ({ ok: false, error: { message: 'unused' } }),
     }))
@@ -252,7 +252,7 @@ describe('--from-url', () => {
   test(
     'timeout produces OutdatedError with E_FETCH_TIMEOUT',
     async () => {
-      mock.module('../src/lib/git', () => ({
+      mock.module('@agents/core/git', () => ({
         fetchSkillFolderHash: async () => null,
         lsRemote: async () => ({ ok: false, error: { message: 'unused' } }),
       }))
@@ -295,7 +295,7 @@ describe('git source', () => {
   test('reports "current" when lsRemote returns matching hash', async () => {
     const HASH = 'e'.repeat(64)
 
-    mock.module('../src/lib/git', () => ({
+    mock.module('@agents/core/git', () => ({
       fetchSkillFolderHash: async () => null,
       lsRemote: async () => ({ ok: true, value: HASH }),
     }))
@@ -318,7 +318,7 @@ describe('git source', () => {
   test('reports "outdated" when lsRemote returns different hash', async () => {
     const REMOTE = 'b'.repeat(64)
 
-    mock.module('../src/lib/git', () => ({
+    mock.module('@agents/core/git', () => ({
       fetchSkillFolderHash: async () => null,
       lsRemote: async () => ({ ok: true, value: REMOTE }),
     }))
@@ -339,7 +339,7 @@ describe('git source', () => {
   })
 
   test('reports "unavailable" when lsRemote fails', async () => {
-    mock.module('../src/lib/git', () => ({
+    mock.module('@agents/core/git', () => ({
       fetchSkillFolderHash: async () => null,
       lsRemote: async () => ({
         ok: false,
@@ -378,7 +378,7 @@ describe('mixed sources', () => {
 
     const GITHUB_HASH = 'b'.repeat(64)
 
-    mock.module('../src/lib/git', () => ({
+    mock.module('@agents/core/git', () => ({
       fetchSkillFolderHash: async () => GITHUB_HASH,
       lsRemote: async () => ({ ok: true, value: 'unused' }),
     }))
