+++ packages/core/package.json
@@ -10,6 +10,9 @@
     "./component/*": "./src/component/*.ts"
   },
   "dependencies": {
+    "@napi-rs/keyring": "^1.2.0",
+    "@octokit/auth-oauth-device": "^7.0.0",
+    "@octokit/core": "^6.0.0",
     "ansis": "^4.0.0",
     "console-table-printer": "^2.15.0",
     "nanospinner": "^1.2.0",
