+++ packages/core/src/source-parser.ts
